requirements.txt
azure-functions
azure-data-tables
python-dateutil

